Move logger to util package
Add smaller interfaces on cmd and imageset for the logger
