+++ pkg/imgpkg/cmd/ui_flags.go
@@ -9,6 +9,11 @@ import (
 	"github.com/spf13/cobra"
 )
 
+type Logger interface {
+	Write(data []byte) (int, error)
+	WriteStr(str string, args ...interface{}) error
+}
+
 type UIFlags struct {
 	TTY            bool
 	Color          bool
